requirements-dev-raw.txt
numpy
jax
jaxlib
matplotlib
msgpack
tqdm
pathos


# test requirements
pytest
absl-py
